version 0.0.2: added feature to simplify polynomial to reduce needed computations
